+++ ApplicationLibCode/UserInterface/RiuSelectionChangedHandler.cpp
@@ -447,14 +447,15 @@ void RiuSelectionChangedHandler::addDepthCurveFromSelectionItem( const RiuSelect
 
         std::vector<double> resultValues = RigDepthResultAccessor::resultValues( casedata,
                                                                                  eclResDef,
-                                                                                 eclipseSelectionItem->m_gridIndex,
+                                                                                 static_cast<int>( eclipseSelectionItem->m_gridIndex ),
                                                                                  eclipseSelectionItem->m_gridLocalCellIndex,
                                                                                  currentTimeStep );
 
-        std::vector<int> kValues = RigDepthResultAccessor::kValues( casedata, eclipseSelectionItem->m_gridIndex );
+        std::vector<int> kValues = RigDepthResultAccessor::kValues( casedata, static_cast<int>( eclipseSelectionItem->m_gridIndex ) );
 
-        std::vector<double> depthValues =
-            RigDepthResultAccessor::depthValues( casedata, eclipseSelectionItem->m_gridLocalCellIndex, eclipseSelectionItem->m_gridIndex );
+        std::vector<double> depthValues = RigDepthResultAccessor::depthValues( casedata,
+                                                                               static_cast<int>( eclipseSelectionItem->m_gridLocalCellIndex ),
+                                                                               static_cast<int>( eclipseSelectionItem->m_gridIndex ) );
 
         CVF_ASSERT( kValues.size() == resultValues.size() );
